pyproject.toml
[project]
name = "finbot"
version = "0.1.0"
description = ""
authors = [{ name = "Bruno Flávio Ferreira", email = "ndninguem9@gmail.com" }]
requires-python = "~=3.10"
dependencies = [
    "langchain-ollama>=0.2.3,<0.3",
    "langchain-community>=0.3.19,<0.4",
    "langgraph>=0.3.5,<0.4",
    "urllib3>= 1.25.3, < 3.0.0",
    "python-dateutil>= 2.8.2",
    "pydantic>= 2",
    "typing-extensions>= 4.7.1",
    "langchain-openai>=0.3.11,<0.4",
    "openapi-client",
    "python-telegram-bot~=22.0",
    "asyncio>=3.4.3,<4",
    "pypdf>=5.4.0,<6",
    "numpy>=2.2.4,<3",
]

[dependency-groups]
dev = [
    "pytest>= 7.2.1",
    "pytest-cov>= 2.8.1",
    "tox>= 3.9.0",
    "flake8>= 4.0.0",
    "types-python-dateutil>= 2.8.19.14",
    "mypy>= 1.5",
]

[tool.uv]

[tool.uv.sources]
openapi-client = { path = "generated", editable = true }

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

